added wind
wind has four parameters

_WaveDir("Wind Direction", Vector)
_WaveSpeed("Wind Speed", float)
_WaveNoise("Wind Noisiness", float)
_WaveScale("Wind Scale", float)

which should be accessible to a global manager that can set all of them.
(Unless there is a way to get a global value in a shader?)

It's based on code given to me by Chris but modified a little bit to
include influence based on distance from the ground and more directional
swaying (as opposed to oscillating 1:1)

I've also categorized this family of shaders as "InsideWrap/," removed
the shadow color parameter from one of the shaders, recolored the cover
to fit the color of the surface and added vertex colors to all of the
plant life (with the exception of sprites which will need to be modified
at some point).