transparent color shader now does billboarding
It's really inefficient but unclear which is the lesser of two evils:
more draw calls on gpu or higher cpu usage? Billboarding totally falls
apart when batching and everything looks like its moving at once :'~(

It's worth disabling your billboardme script in prefab. If you don't
want to you can always go into the material and toggle the billboard
effect in the material. Comment out the tag section
"DisableBatching"="True" and turn off billboard if performance is crappy
in comparison to script.