singing pass
added the extra stems to the project (possibly to be used in the future
for fading in different parts dynamically), as well as the world
singing and player singing tracks, changed the audio manager so it
fades in the singing track when you hit the button, and in the player
singing script i made it fade down on release.

I think this approach might work if I update the actual singing tracks,
though another cool approach might be just having a bunch of clips of
the singing and the player hits one note at a time. Even though this
will probably clash with the music a little bit, it might seem more
responsive since right now I think the feels like a canned track that
fades up.

Also, Murray’s FFT approach may work too now that I split up the
tracks, since the pitch could be grabbed from one of the non-drum
tracks.

woo!